refactor feature modal and plot out of table component file
